+++ src/ComputeSharp.Graphics/Buffers/Abstract/Texture2D{T}.cs
@@ -127,7 +127,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         /// <param name="y">The vertical offset in the source texture.</param>
         /// <param name="width">The width of the memory area to copy.</param>
         /// <param name="height">The height of the memory area to copy.</param>
-        internal void GetData(ref T destination, nint size, int x, int y, int width, int height)
+        internal void GetData(ref T destination, int size, int x, int y, int width, int height)
         {
             GraphicsDevice.ThrowIfDisposed();
 
@@ -199,7 +199,7 @@ namespace ComputeSharp.Graphics.Buffers.Abstract
         /// <param name="y">The vertical offset in the destination texture.</param>
         /// <param name="width">The width of the memory area to write to.</param>
         /// <param name="height">The height of the memory area to write to.</param>
-        internal void SetData(ref T source, nint size, int x, int y, int width, int height)
+        internal void SetData(ref T source, int size, int x, int y, int width, int height)
         {
             GraphicsDevice.ThrowIfDisposed();
